Create qSegmenter plugin
