+++ plugins/core/Standard/qSegmenter/src/CMakeLists.txt
@@ -2,5 +2,5 @@
 target_sources( ${PROJECT_NAME}
 	PRIVATE
 		${CMAKE_CURRENT_LIST_DIR}/ActionA.cpp
-		${CMAKE_CURRENT_LIST_DIR}/ExamplePlugin.cpp
+		${CMAKE_CURRENT_LIST_DIR}/qSegmenter.cpp
 )
