+++ plugins/example/CMakeLists.txt
@@ -1,4 +1,4 @@
 
 add_subdirectory( ${CMAKE_CURRENT_SOURCE_DIR}/ExampleGLPlugin )
 add_subdirectory( ${CMAKE_CURRENT_SOURCE_DIR}/ExampleIOPlugin )
-add_subdirectory( ${CMAKE_CURRENT_SOURCE_DIR}/ExamplePlugin )
+# add_subdirectory( ${CMAKE_CURRENT_SOURCE_DIR}/ExamplePlugin )
