--- plugins/example/ExamplePlugin/ExamplePlugin.qrc
@@ -1,6 +1,0 @@
-<RCC>
-  <qresource prefix="/CC/plugin/ExamplePlugin" >
-    <file>images/icon.png</file>
-    <file>info.json</file>
-  </qresource>
-</RCC>
