--- plugins/example/ExamplePlugin/info.json
@@ -1,34 +1,0 @@
-{
-	"type" : "Standard",
-	"name" : "Example (Standard Plugin)",
-	"icon" : ":/CC/plugin/ExamplePlugin/images/icon.png",
-	"description": "This is a description of the marvelous Example plugin. It does nothing.",
-	"authors" : [
-		{
-			"name" : "Daniel Girardeau-Montaut",
-			"email" : "daniel.girardeau@gmail.com"
-		}
-	],
-	"maintainers" : [
-		{
-			"name" : "Andy Maloney",
-			"email" : "asmaloney@gmail.com"
-		},
-		{
-			"name" : "Example NoEmail"
-		}
-	],
-	"references" : [
-		{
-			"text" : "The unsuccessful self-treatment of a case of “writer's block”",
-			"url" : "https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1311997/"
-		},
-		{
-			"text" : "Sword swallowing and its side effects",
-			"url" : "http://www.bmj.com/content/333/7582/1285"
-		},
-		{
-			"text" : "I thought of creating this wonderful plugin while I was hiking up &#x26f0; Mont Blanc"
-		}
-	]
-}
